requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
jinja2>=3.1.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Web Scraping
playwright>=1.41.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Scheduling
apscheduler>=3.10.0

# CLI
click>=8.1.0
rich>=13.7.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Platform SDKs (optional)
# lazop-sdk  # Lazada Open Platform
# pyshopee   # Shopee (unofficial)

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
